third_party/WebKit/LayoutTests/fast/dom/Element/setAttributeNode-case-insensitivity-expected.txt

Test for Bugzilla bug: 90341: createAttribute/setAttributeNode does not properly normalize case.
 
This test verifies that the setAttributeNode() API checks for existing attributes case-insensitively. Thus the value of an existing attribute with the same name but in a different case should get replaced by the new value specified via the setAttributeNode() method.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
PASS test.getAttribute('style') is test.getAttribute('STYLE')
PASS test.getAttributeNode('style').value is test.getAttributeNode('STYLE').value
